# apps/api/api/openapi.yaml
openapi: 3.0.3
info:
  title: Majadu API
  version: 0.1.0
  description: |
    Backend Go untuk Majadu (badminton session operations).
    Kontrak RESTful — pengganti PostgREST RPC. Frontend query layer
    akan menyesuaikan dengan kontrak ini.
  license:
    name: Private
servers:
  - url: https://api.qouver.com/majadu
    description: Production (base path /majadu)
  - url: https://api.qouver.com/majadu
    description: Dev/staging (base path /majadu)
...
paths:
...
  /sessions:
    get:
      tags: [Sessions]
      summary: List metadata session
      operationId: listSessions
      responses:
        "200":
          description: Daftar metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SessionMeta" }
    post:
      tags: [Sessions]
      summary: Buat session baru
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SessionDraft" }
      responses:
        "201":
          description: Session dibuat
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: Validasi gagal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /sessions/{id}/players/{playerId}:
    patch:
      tags: [Players]
      summary: Rename pemain dalam session
      operationId: renamePlayer
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/PlayerId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RenamePlayer" }
      responses:
        "200":
          description: Pemain di-rename
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
  /sessions/{id}/absent:
    put:
      tags: [Players]
      summary: Set daftar pemain absent
      operationId: setAbsent
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Absent" }
      responses:
        "200":
          description: Absent tersimpan
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }

  # ── Swaps ─────────────────────────────────────────────────────────────
  /sessions/{id}/swaps/players:
    post:
      tags: [Sessions]
      summary: Tukar dua pemain antar game
      operationId: swapPlayers
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Swap" }
      responses:
        "200":
          description: Swap diterapkan
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
  /sessions/{id}/swaps/teams:
    post:
      tags: [Sessions]
      summary: Tukar dua team antar game
      operationId: swapTeams
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Swap" }
      responses:
        "200":
          description: Swap diterapkan
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
  /sessions/{id}/swaps/slots:
    post:
      tags: [Sessions]
      summary: Tukar dua slot game
      operationId: swapSlots
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Swap" }
      responses:
        "200":
          description: Swap diterapkan
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }

  # ── Players (registry) ────────────────────────────────────────────────
...
components:
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: share_code atau uuid session
...
    PlayerId:
      name: playerId
      in: path
      required: true
      schema: { type: string }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
      description: Version session dalam format `v{n}`. Opsional — kalau ada, server tolak kalau tidak cocok (optimistic concurrency).
  headers:
    ETag:
      description: Version session saat ini (`v{n}`) — untuk If-Match berikutnya
      schema: { type: string }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [not_found, conflict, validation_error, unauthorized, service_unavailable, internal, database_error]
            message: { type: string }
      required: [error]
    SessionMeta:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        date: { type: string, format: date }
        player_count: { type: integer }
        total_games: { type: integer }
        locked: { type: boolean }
      required: [id, title, date, player_count, total_games, locked]
    SessionDraft:
      type: object
      description: Data session untuk create (server generate id + version)
      required: [title, date, courts, players]
      properties:
        title: { type: string }
        date: { type: string, format: date }
        courts: { type: integer, minimum: 1, maximum: 6 }
        sessionStart: { type: string, example: "09:00" }
        slotMinutes: { type: integer, default: 20 }
        courtTimes:
          type: array
          items:
            type: object
            properties:
              start: { type: string }
              end: { type: string }
        players:
          type: array
          minItems: 4
          maxItems: 40
          items: { $ref: "#/components/schemas/Player" }
        fixMatches:
          type: array
          items: { $ref: "#/components/schemas/FixMatch" }
        schedule:
          type: array
          items: { $ref: "#/components/schemas/ScheduleSlot" }
    Session:
      type: object
      description: Session lengkap (mirror CloudSnapshot frontend)
      properties:
        version: { type: integer }
        session: { $ref: "#/components/schemas/SessionConfig" }
        players: { type: array, items: { $ref: "#/components/schemas/Player" } }
        fixMatches: { type: array, items: { $ref: "#/components/schemas/FixMatch" } }
        schedule: { type: array, items: { $ref: "#/components/schemas/ScheduleSlot" } }
        playedGames: { type: array, items: { type: string } }
        gameScores:
          type: object
          additionalProperties: { $ref: "#/components/schemas/Score" }
        absentPlayers: { type: array, items: { type: string } }
    SessionConfig:
      type: object
      properties:
        title: { type: string }
        date: { type: string, format: date }
        courts: { type: integer }
        sessionStart: { type: string }
        slotMinutes: { type: integer }
        courtTimes: { type: array, items: { type: object } }
        playerCount: { type: integer }
        courtNames: { type: array, items: { type: string } }
        locked: { type: boolean }
    Player:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        gender: { type: string, enum: [M, F] }
        tier: { type: integer, minimum: 1, maximum: 4 }
      required: [id, name]
    FixMatch:
      type: object
      properties:
        id: { type: string }
        slots:
          type: array
          items: { type: [string, "null"] }
          minItems: 4
          maxItems: 4
    ScheduleSlot:
      type: object
      properties:
        slot: { type: integer }
        court: { type: integer }
        teamA:
          type: array
          items: { type: string }
          minItems: 2
          maxItems: 2
        teamB:
          type: array
          items: { type: string }
          minItems: 2
          maxItems: 2
...
    Score:
      type: object
      required: [scoreA, scoreB]
      properties:
        scoreA: { type: integer, minimum: 0, maximum: 99 }
        scoreB: { type: integer, minimum: 0, maximum: 99 }
    RenamePlayer:
      type: object
      required: [name]
      properties:
        name: { type: string }
    Absent:
      type: object
      required: [playerIds]
      properties:
        playerIds: { type: array, items: { type: string } }
    Swap:
      type: object
      description: Dua target yang ditukar (struktur tergantung jenis swap)
      required: [targetA, targetB]
      properties:
        targetA: { $ref: "#/components/schemas/SwapTarget" }
        targetB: { $ref: "#/components/schemas/SwapTarget" }
    SwapTarget:
      type: object
      properties:
        slot: { type: integer }
        court: { type: integer }
        team: { type: string, enum: [A, B] }
        position: { type: integer }
...